+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 2.4.27
 ### Added
 * Ability to filter by artwork size for banners, heroes, and icons.
-
+* Ability to sort exceptions by date, old title, and new title ([issue 343](https://github.com/SteamGridDB/steam-rom-manager/issues/343))
 ### Changed
 * Hide API options for SteamGridDB if it isn't selected as an image provider.